action-planner/pyproject.toml
[project]
name = "action-planner"
version = "1.0.0"
description = "Autonomous Action Planning Agent - Converts feedback insights into trackable action items"
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
keywords = ["ai", "agents", "action-planning", "jira", "a2a"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    # Agent Framework
    "agent-framework==1.0.0b251209",
    "agent-framework-ag-ui==1.0.0b251209",
    "agent-framework-devui==1.0.0b251209",
    "agent-framework-azure-ai==1.0.0b251209",

    # Azure AI
    "azure-ai-agents==1.2.0b5",
    "azure-ai-projects==2.0.0b2",
    "azure-identity>=1.15.0",

    # Web Framework
    "fastapi>=0.115.0",
    "uvicorn>=0.30.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "httpx>=0.27.0",

    # Ticketing Systems Integration
    "jira>=3.5.0",  # Atlassian Jira

    # OpenTelemetry (optional)
    "opentelemetry-api>=1.22.0",
    "opentelemetry-sdk>=1.22.0",
    "opentelemetry-instrumentation-fastapi>=0.43b0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pyright>=1.1.0",
    "ruff>=0.4.0",
]

[project.scripts]
action-planner = "actionplanner.__main__:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["actionplanner"]

[tool.pyright]
pythonVersion = "3.11"
typeCheckingMode = "basic"
include = ["actionplanner"]

[tool.ruff]
line-length = 120
target-version = "py311"

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
